+++ src/src/java/controller/SystemGroupBean.java
@@ -68,7 +68,7 @@ public class SystemGroupBean implements Serializable {
     }
 
     public int getPageCount() {
-        int numOfRowOnPage = 5;
+        double numOfRowOnPage = 5;
         List<SystemGroup> list = this.getDao().readList();
         pageCount = (int) Math.ceil(list.size() / numOfRowOnPage);
         return pageCount;
